refactor: replace `scrollIntoView` with spring animation
Previously in order to scroll down to the next project
`scrollIntoView` was used. There are no options to customise the
easing however, and it has to be polyfilled. This commit replaces
it with react spring instead.
Note that we do lose the ability to maintain previous scroll
position on forward/back because of `position: fixed`, but since
the site is a linear journey it shouldnt be a major problem